Move Period class to a shared location to be also used by KPIs
